+++ components/user/EntryTypeGraph.jsx
@@ -1,7 +1,7 @@
 import { Col, Row } from "react-bootstrap"
-import { BarChart, Tooltip, Bar, XAxis, Pie, PieChart, Cell, Legend, YAxis, ResponsiveContainer, LineChart, Line, CartesianGrid } from "recharts"
+import { Label, BarChart, Tooltip, Bar, XAxis, Pie, PieChart, Cell, Legend, YAxis, ResponsiveContainer, LineChart, Line, CartesianGrid } from "recharts"
 import { colourList } from "../../public/helpers/frontend"
-import { msToStr } from "../../public/helpers/frontendConverters"
+import { msToStr, processLinePlotData } from "../../public/helpers/frontendConverters"
 
 const EntryTypeGraph = ({ data }) => {
   const barChartData = []
@@ -54,17 +54,7 @@ const EntryTypeGraph = ({ data }) => {
     percOfTotal: data.wt
   })
 
-  const netherDistData = []
-  const sortedData = [...data.nd].sort((a, b) => a - b);
-  const start = sortedData[0]
-  const end = sortedData[Math.trunc(sortedData.length * 0.98)]
-  const step = 1
-  const dataRange = Array.from(Array(Math.ceil((end - start) / step)).keys(), (i) => start + i * step)
-  dataRange.forEach((num, idx) => {
-    const index = sortedData.findIndex((element) => element >= num)
-    const count1 = index !== -1 ? index : sortedData.length
-    netherDistData.push({time: num, count: count1})
-})
+  const netherDistData = processLinePlotData(data.nd, 2000)
 
 
 
@@ -151,6 +141,22 @@ const EntryTypeGraph = ({ data }) => {
       </Row>
       <Row style={{ width: "100%" }}>
         <Col style={{ height: "300px" }} className="d-flex flex-column col-md-6 col-sm-12">
+          <h1>Enter Time Distribution</h1>
+          <ResponsiveContainer>
+            <LineChart width={500} height={250} data={netherDistData}>
+              <CartesianGrid strokeDasharray="3 3" />
+              <XAxis dataKey="time" tickFormatter={value => [msToStr(value), ""]} stroke="#b2b2b2" >
+                <Label value="X (time)" offset={0} position="insideBottom" style={{ textAnchor: 'middle', fill: '#b2b2b2' }}/>
+              </XAxis>
+              <YAxis stroke="#b2b2b2" >
+                <Label value="Sub-X enter count" offset={0} position="insideLeft" angle={-90} style={{ textAnchor: 'middle', fill: '#b2b2b2' }} />
+              </YAxis>
+              <Tooltip />
+              <Line type="monotone" dataKey="count" stroke="#8884d8" strokeWidth={3} dot={{ r: 2 }}/>
+            </LineChart>
+          </ResponsiveContainer>
+        </Col>
+        <Col style={{ height: "300px" }} className="d-flex flex-column col-md-6 col-sm-12">
           <h1>Playtime Breakdown</h1>
           <ResponsiveContainer>
             <PieChart width={300} height={250} className="mx-auto">
@@ -168,21 +174,8 @@ const EntryTypeGraph = ({ data }) => {
                 ))}
               </Pie>
               <Legend layout="horizontal" verticalAlign="bottom" align="right" />
-              <Tooltip />
+              <Tooltip separator="" formatter={value => [msToStr(value), ""]} cursor={false} itemStyle={{ color: "#000000" }} labelStyle={{ color: "#000000" }} />
             </PieChart>
-          </ResponsiveContainer>
-        </Col>
-        <Col style={{ height: "300px" }} className="d-flex flex-column col-md-6 col-sm-12">
-          <h1>Enter Time Distribution</h1>
-          <ResponsiveContainer>
-            <LineChart width={300} height={250} data={netherDistData}>
-              <CartesianGrid strokeDasharray="3 3" />
-              <XAxis dataKey="time" />
-              <YAxis />
-              <Tooltip />
-              <Legend />
-              <Line type="monotone" dataKey="count" stroke="#8884d8" strokeWidth={3} dot={{ r: 2 }}/>
-            </LineChart>
           </ResponsiveContainer>
         </Col>
       </Row>
